+++ internal/framework/volume.go
@@ -9,18 +9,22 @@ func (h *Helper) deleteVolumes() error {
 	opts := volumes.ListOpts{TenantID: h.Spec.Openstack.Project.ID}
 	volumes, err := h.Openstack.ListVolumes(opts)
 	if err != nil {
-		log.Errorf("framework: failed to list openstack volumes(%v)", err)
+		log.Errorf("openstack: failed to list openstack volumes(%v)", err)
 		return err
 	}
 
 	for _, volume := range volumes {
+		if volume.TenantID != h.Spec.Openstack.Project.ID {
+			continue
+		}
+
 		err = h.Openstack.DeleteVolume(volume.ID)
 		if err != nil {
-			log.Errorf("framework: failed to delete openstack volume %s(%v)", volume.Name, err)
+			log.Errorf("openstack: failed to delete openstack volume %s(%v)", volume.Name, err)
 			return err
 		}
 
-		log.Infof("framework: deleted openstack volume %s", volume.Name)
+		log.Infof("openstack: deleted openstack volume %s", volume.Name)
 	}
 
 	return nil
